Extract more code from NameResolutionAndTypeInference
