+++ Source/DafnyCore/Resolver/NameResolutionAndTypeInference/XConstraint.cs
@@ -60,47 +60,47 @@ public class XConstraint {
     bool satisfied;
     switch (ConstraintName) {
       case "Assignable": {
-        Contract.Assert(t == t.Normalize());  // it's already been normalized above
-        var u = Types[1].NormalizeExpand();
-        if (CheckTypeInferenceVisitor.IsDetermined(t) &&
-            (fullstrength
-             || !ProxyWithNoSubTypeConstraint(u, resolver)
-             || (u is TypeProxy
-                 && Types[0].NormalizeExpandKeepConstraints() is var t0constrained
-                 && (t0constrained.IsNonNullRefType || t0constrained.AsSubsetType != null)
-                 && resolver.HasApplicableNullableRefTypeConstraint(new HashSet<TypeProxy>() { (TypeProxy)u })))) {
-          // This is the best case.  We convert Assignable(t, u) to the subtype constraint base(t) :> u.
-          if (CheckTypeInferenceVisitor.IsDetermined(u) && t.IsSubtypeOf(u, false, true) && t.IsRefType) {
-            // But we also allow cases where the rhs is a proper supertype of the lhs, and let the verifier
-            // determine whether the rhs is provably an instance of the lhs.
-            resolver.ConstrainAssignable((NonProxyType)u, (NonProxyType)t, errorMsg, out moreXConstraints, fullstrength);
-          } else {
+          Contract.Assert(t == t.Normalize());  // it's already been normalized above
+          var u = Types[1].NormalizeExpand();
+          if (CheckTypeInferenceVisitor.IsDetermined(t) &&
+              (fullstrength
+               || !ProxyWithNoSubTypeConstraint(u, resolver)
+               || (u is TypeProxy
+                   && Types[0].NormalizeExpandKeepConstraints() is var t0constrained
+                   && (t0constrained.IsNonNullRefType || t0constrained.AsSubsetType != null)
+                   && resolver.HasApplicableNullableRefTypeConstraint(new HashSet<TypeProxy>() { (TypeProxy)u })))) {
+            // This is the best case.  We convert Assignable(t, u) to the subtype constraint base(t) :> u.
+            if (CheckTypeInferenceVisitor.IsDetermined(u) && t.IsSubtypeOf(u, false, true) && t.IsRefType) {
+              // But we also allow cases where the rhs is a proper supertype of the lhs, and let the verifier
+              // determine whether the rhs is provably an instance of the lhs.
+              resolver.ConstrainAssignable((NonProxyType)u, (NonProxyType)t, errorMsg, out moreXConstraints, fullstrength);
+            } else {
+              resolver.ConstrainAssignable((NonProxyType)t, u, errorMsg, out moreXConstraints, fullstrength);
+            }
+            convertedIntoOtherTypeConstraints = true;
+            return true;
+          } else if (u.IsTypeParameter) {
+            // we need the constraint base(t) :> u, which for a type parameter t can happen iff t :> u
+            resolver.ConstrainSubtypeRelation(t, u, errorMsg);
+            convertedIntoOtherTypeConstraints = true;
+            return true;
+          } else if (Type.FromSameHead(t, u, out var tUp, out var uUp)) {
+            resolver.ConstrainAssignableTypeArgs(tUp, tUp.TypeArgs, uUp.TypeArgs, errorMsg, out moreXConstraints);
+            return true;
+          } else if (fullstrength && t is NonProxyType) {
+            // We convert Assignable(t, u) to the subtype constraint base(t) :> u.
             resolver.ConstrainAssignable((NonProxyType)t, u, errorMsg, out moreXConstraints, fullstrength);
+            convertedIntoOtherTypeConstraints = true;
+            return true;
+          } else if (fullstrength && u is NonProxyType) {
+            // We're willing to change "base(t) :> u" to the stronger constraint "t :> u" for the sake of making progress.
+            resolver.ConstrainSubtypeRelation(t, u, errorMsg);
+            convertedIntoOtherTypeConstraints = true;
+            return true;
           }
-          convertedIntoOtherTypeConstraints = true;
-          return true;
-        } else if (u.IsTypeParameter) {
-          // we need the constraint base(t) :> u, which for a type parameter t can happen iff t :> u
-          resolver.ConstrainSubtypeRelation(t, u, errorMsg);
-          convertedIntoOtherTypeConstraints = true;
-          return true;
-        } else if (Type.FromSameHead(t, u, out var tUp, out var uUp)) {
-          resolver.ConstrainAssignableTypeArgs(tUp, tUp.TypeArgs, uUp.TypeArgs, errorMsg, out moreXConstraints);
-          return true;
-        } else if (fullstrength && t is NonProxyType) {
-          // We convert Assignable(t, u) to the subtype constraint base(t) :> u.
-          resolver.ConstrainAssignable((NonProxyType)t, u, errorMsg, out moreXConstraints, fullstrength);
-          convertedIntoOtherTypeConstraints = true;
-          return true;
-        } else if (fullstrength && u is NonProxyType) {
-          // We're willing to change "base(t) :> u" to the stronger constraint "t :> u" for the sake of making progress.
-          resolver.ConstrainSubtypeRelation(t, u, errorMsg);
-          convertedIntoOtherTypeConstraints = true;
-          return true;
+          // There's not enough information to say anything
+          return false;
         }
-        // There's not enough information to say anything
-        return false;
-      }
       case "NumericType":
         satisfied = t.IsNumericBased();
         break;
@@ -123,13 +123,13 @@ public class XConstraint {
         satisfied = t.IsNumericBased() || t.IsBitVectorType || t.IsBigOrdinalType || t.IsCharType || t is SetType || t is MultiSetType;
         break;
       case "RankOrderable": {
-        var u = Types[1].NormalizeExpand();
-        if (u is TypeProxy) {
-          return false;  // not enough information
+          var u = Types[1].NormalizeExpand();
+          if (u is TypeProxy) {
+            return false;  // not enough information
+          }
+          satisfied = (t.IsIndDatatype || t.IsTypeParameter) && u.IsIndDatatype;
+          break;
         }
-        satisfied = (t.IsIndDatatype || t.IsTypeParameter) && u.IsIndDatatype;
-        break;
-      }
       case "Plussable":
         satisfied = t.IsNumericBased() || t.IsBitVectorType || t.IsBigOrdinalType || t.IsCharType || t is SeqType || t is SetType || t is MultiSetType || t is MapType;
         break;
@@ -234,47 +234,47 @@ public class XConstraint {
         }
         break;
       case "Innable": {
-        var elementType = FindCollectionType(resolver.Options, t, true, new HashSet<TypeProxy>()) ?? FindCollectionType(resolver.Options, t, false, new HashSet<TypeProxy>());
-        if (elementType != null) {
-          var u = Types[1];  // note, it's okay if "u" is a TypeProxy
-          resolver.AddXConstraint(this.tok, "Equatable", elementType, u, new TypeConstraint.ErrorMsgWithBase(errorMsg, "expecting element type to be assignable to {1} (got {0})", u, elementType));
-          moreXConstraints = true;
-          return true;
-        }
-        if (t is TypeProxy) {
-          return false;  // not enough information to do anything
-        }
-        satisfied = false;
-        break;
-      }
-      case "SeqUpdatable": {
-        var xcWithExprs = (Resolver.XConstraintWithExprs)this;
-        var index = xcWithExprs.Exprs[0];
-        var value = xcWithExprs.Exprs[1];
-        if (t is SeqType) {
-          var s = (SeqType)t;
-          resolver.ConstrainToIntegerType(index, true, "sequence update requires integer- or bitvector-based index (got {0})");
-          resolver.ConstrainSubtypeRelation(s.Arg, value.Type, value, "sequence update requires the value to have the element type of the sequence (got {0})", value.Type);
-        } else if (t is MapType) {
-          var s = (MapType)t;
-          if (s.Finite) {
-            resolver.ConstrainSubtypeRelation(s.Domain, index.Type, index, "map update requires domain element to be of type {0} (got {1})", s.Domain, index.Type);
-            resolver.ConstrainSubtypeRelation(s.Range, value.Type, value, "map update requires the value to have the range type {0} (got {1})", s.Range, value.Type);
-          } else {
-            resolver.ConstrainSubtypeRelation(s.Domain, index.Type, index, "imap update requires domain element to be of type {0} (got {1})", s.Domain, index.Type);
-            resolver.ConstrainSubtypeRelation(s.Range, value.Type, value, "imap update requires the value to have the range type {0} (got {1})", s.Range, value.Type);
+          var elementType = FindCollectionType(resolver.Options, t, true, new HashSet<TypeProxy>()) ?? FindCollectionType(resolver.Options, t, false, new HashSet<TypeProxy>());
+          if (elementType != null) {
+            var u = Types[1];  // note, it's okay if "u" is a TypeProxy
+            resolver.AddXConstraint(this.tok, "Equatable", elementType, u, new TypeConstraint.ErrorMsgWithBase(errorMsg, "expecting element type to be assignable to {1} (got {0})", u, elementType));
+            moreXConstraints = true;
+            return true;
           }
-        } else if (t is MultiSetType) {
-          var s = (MultiSetType)t;
-          resolver.ConstrainSubtypeRelation(s.Arg, index.Type, index, "multiset update requires domain element to be of type {0} (got {1})", s.Arg, index.Type);
-          resolver.ConstrainToIntegerType(value, false, "multiset update requires integer-based numeric value (got {0})");
-        } else {
+          if (t is TypeProxy) {
+            return false;  // not enough information to do anything
+          }
           satisfied = false;
           break;
         }
-        convertedIntoOtherTypeConstraints = true;
-        return true;
-      }
+      case "SeqUpdatable": {
+          var xcWithExprs = (XConstraintWithExprs)this;
+          var index = xcWithExprs.Exprs[0];
+          var value = xcWithExprs.Exprs[1];
+          if (t is SeqType) {
+            var s = (SeqType)t;
+            resolver.ConstrainToIntegerType(index, true, "sequence update requires integer- or bitvector-based index (got {0})");
+            resolver.ConstrainSubtypeRelation(s.Arg, value.Type, value, "sequence update requires the value to have the element type of the sequence (got {0})", value.Type);
+          } else if (t is MapType) {
+            var s = (MapType)t;
+            if (s.Finite) {
+              resolver.ConstrainSubtypeRelation(s.Domain, index.Type, index, "map update requires domain element to be of type {0} (got {1})", s.Domain, index.Type);
+              resolver.ConstrainSubtypeRelation(s.Range, value.Type, value, "map update requires the value to have the range type {0} (got {1})", s.Range, value.Type);
+            } else {
+              resolver.ConstrainSubtypeRelation(s.Domain, index.Type, index, "imap update requires domain element to be of type {0} (got {1})", s.Domain, index.Type);
+              resolver.ConstrainSubtypeRelation(s.Range, value.Type, value, "imap update requires the value to have the range type {0} (got {1})", s.Range, value.Type);
+            }
+          } else if (t is MultiSetType) {
+            var s = (MultiSetType)t;
+            resolver.ConstrainSubtypeRelation(s.Arg, index.Type, index, "multiset update requires domain element to be of type {0} (got {1})", s.Arg, index.Type);
+            resolver.ConstrainToIntegerType(value, false, "multiset update requires integer-based numeric value (got {0})");
+          } else {
+            satisfied = false;
+            break;
+          }
+          convertedIntoOtherTypeConstraints = true;
+          return true;
+        }
       case "ContainerIndex":
         // The semantics of this XConstraint is that *if* the head is seq/array/map/multiset, then its element/domain type must a supertype of "u"
         Type indexType;
@@ -314,187 +314,187 @@ public class XConstraint {
         convertedIntoOtherTypeConstraints = true;
         return true;
       case "Equatable": {
-        t = Types[0].NormalizeExpandKeepConstraints();
-        var u = Types[1].NormalizeExpandKeepConstraints();
-        if (object.ReferenceEquals(t, u)) {
-          return true;
-        }
-        if (t is TypeProxy && u is TypeProxy) {
-          return false;  // not enough information to do anything sensible
-        } else if (t is TypeProxy || u is TypeProxy) {
-          TypeProxy proxy;
-          Type other;
-          if (t is TypeProxy) {
-            proxy = (TypeProxy)t;
-            other = u;
-          } else {
-            proxy = (TypeProxy)u;
-            other = t;
-          }
-          if (other.IsNumericBased() || other.IsBitVectorType || other.IsBigOrdinalType) {
-            resolver.ConstrainSubtypeRelation(other.NormalizeExpand(), proxy, errorMsg, true);
-            convertedIntoOtherTypeConstraints = true;
+          t = Types[0].NormalizeExpandKeepConstraints();
+          var u = Types[1].NormalizeExpandKeepConstraints();
+          if (object.ReferenceEquals(t, u)) {
             return true;
-          } else if (fullstrength) {
-            // the following is rather aggressive
-            if (Resolver.TypeConstraintsIncludeProxy(other, proxy)) {
-              return false;
+          }
+          if (t is TypeProxy && u is TypeProxy) {
+            return false;  // not enough information to do anything sensible
+          } else if (t is TypeProxy || u is TypeProxy) {
+            TypeProxy proxy;
+            Type other;
+            if (t is TypeProxy) {
+              proxy = (TypeProxy)t;
+              other = u;
             } else {
-              if (other.IsRefType && resolver.HasApplicableNullableRefTypeConstraint_SubDirection(proxy)) {
-                other = other.NormalizeExpand();  // shave off all constraints
-              }
-              satisfied = resolver.AssignProxyAndHandleItsConstraints(proxy, other, true);
-              convertedIntoOtherTypeConstraints = true;
-              break;
+              proxy = (TypeProxy)u;
+              other = t;
             }
-          } else {
-            return false;  // not enough information
+            if (other.IsNumericBased() || other.IsBitVectorType || other.IsBigOrdinalType) {
+              resolver.ConstrainSubtypeRelation(other.NormalizeExpand(), proxy, errorMsg, true);
+              convertedIntoOtherTypeConstraints = true;
+              return true;
+            } else if (fullstrength) {
+              // the following is rather aggressive
+              if (Resolver.TypeConstraintsIncludeProxy(other, proxy)) {
+                return false;
+              } else {
+                if (other.IsRefType && resolver.HasApplicableNullableRefTypeConstraint_SubDirection(proxy)) {
+                  other = other.NormalizeExpand();  // shave off all constraints
+                }
+                satisfied = resolver.AssignProxyAndHandleItsConstraints(proxy, other, true);
+                convertedIntoOtherTypeConstraints = true;
+                break;
+              }
+            } else {
+              return false;  // not enough information
+            }
           }
-        }
 
-        satisfied = Type.FromSameHead_Subtype(t, u, out var a, out var b);
-        if (satisfied) {
-          Contract.Assert(a.TypeArgs.Count == b.TypeArgs.Count);
-          var cl = a is UserDefinedType ? ((UserDefinedType)a).ResolvedClass : null;
-          for (int i = 0; i < a.TypeArgs.Count; i++) {
-            resolver.AllXConstraints.Add(new Resolver.XConstraint_EquatableArg(tok,
-              a.TypeArgs[i], b.TypeArgs[i],
-              a is CollectionType || (cl != null && cl.TypeArgs[i].Variance != TypeParameter.TPVariance.Non),
-              a.IsRefType,
-              errorMsg));
-            moreXConstraints = true;
+          satisfied = Type.FromSameHead_Subtype(t, u, out var a, out var b);
+          if (satisfied) {
+            Contract.Assert(a.TypeArgs.Count == b.TypeArgs.Count);
+            var cl = a is UserDefinedType ? ((UserDefinedType)a).ResolvedClass : null;
+            for (int i = 0; i < a.TypeArgs.Count; i++) {
+              resolver.AllXConstraints.Add(new XConstraint_EquatableArg(tok,
+                a.TypeArgs[i], b.TypeArgs[i],
+                a is CollectionType || (cl != null && cl.TypeArgs[i].Variance != TypeParameter.TPVariance.Non),
+                a.IsRefType,
+                errorMsg));
+              moreXConstraints = true;
+            }
           }
+          break;
         }
-        break;
-      }
       case "EquatableArg": {
-        t = Types[0].NormalizeExpandKeepConstraints();
-        var u = Types[1].NormalizeExpandKeepConstraints();
-        var moreExactThis = (Resolver.XConstraint_EquatableArg)this;
-        if (t is TypeProxy && u is TypeProxy) {
-          return false;  // not enough information to do anything sensible
-        } else if (t is TypeProxy || u is TypeProxy) {
-          TypeProxy proxy;
-          Type other;
-          if (t is TypeProxy) {
-            proxy = (TypeProxy)t;
-            other = u;
-          } else {
-            proxy = (TypeProxy)u;
-            other = t;
-          }
-          if (other.IsNumericBased() || other.IsBitVectorType || other.IsBigOrdinalType) {
-            resolver.ConstrainSubtypeRelation(other.NormalizeExpand(), proxy, errorMsg, true);
-            convertedIntoOtherTypeConstraints = true;
-            return true;
-          } else if (fullstrength) {
-            // the following is rather aggressive
-            if (Resolver.TypeConstraintsIncludeProxy(other, proxy)) {
-              return false;
+          t = Types[0].NormalizeExpandKeepConstraints();
+          var u = Types[1].NormalizeExpandKeepConstraints();
+          var moreExactThis = (XConstraint_EquatableArg)this;
+          if (t is TypeProxy && u is TypeProxy) {
+            return false;  // not enough information to do anything sensible
+          } else if (t is TypeProxy || u is TypeProxy) {
+            TypeProxy proxy;
+            Type other;
+            if (t is TypeProxy) {
+              proxy = (TypeProxy)t;
+              other = u;
             } else {
-              if (other.IsRefType && resolver.HasApplicableNullableRefTypeConstraint_SubDirection(proxy)) {
-                other = other.NormalizeExpand();  // shave off all constraints
-              }
-              satisfied = resolver.AssignProxyAndHandleItsConstraints(proxy, other, true);
-              convertedIntoOtherTypeConstraints = true;
-              break;
+              proxy = (TypeProxy)u;
+              other = t;
             }
-          } else {
-            return false;  // not enough information
+            if (other.IsNumericBased() || other.IsBitVectorType || other.IsBigOrdinalType) {
+              resolver.ConstrainSubtypeRelation(other.NormalizeExpand(), proxy, errorMsg, true);
+              convertedIntoOtherTypeConstraints = true;
+              return true;
+            } else if (fullstrength) {
+              // the following is rather aggressive
+              if (Resolver.TypeConstraintsIncludeProxy(other, proxy)) {
+                return false;
+              } else {
+                if (other.IsRefType && resolver.HasApplicableNullableRefTypeConstraint_SubDirection(proxy)) {
+                  other = other.NormalizeExpand();  // shave off all constraints
+                }
+                satisfied = resolver.AssignProxyAndHandleItsConstraints(proxy, other, true);
+                convertedIntoOtherTypeConstraints = true;
+                break;
+              }
+            } else {
+              return false;  // not enough information
+            }
           }
-        }
-        if (moreExactThis.TreatTypeParamAsWild && (t.IsTypeParameter || u.IsTypeParameter || t.IsAbstractType || u.IsAbstractType)) {
-          return true;
-        } else if (!moreExactThis.AllowSuperSub) {
-          resolver.ConstrainSubtypeRelation_Equal(t, u, errorMsg);
-          convertedIntoOtherTypeConstraints = true;
-          return true;
-        }
+          if (moreExactThis.TreatTypeParamAsWild && (t.IsTypeParameter || u.IsTypeParameter || t.IsAbstractType || u.IsAbstractType)) {
+            return true;
+          } else if (!moreExactThis.AllowSuperSub) {
+            resolver.ConstrainSubtypeRelation_Equal(t, u, errorMsg);
+            convertedIntoOtherTypeConstraints = true;
+            return true;
+          }
 
-        // okay if t<:u or u<:t (this makes type inference more manageable, though it is more liberal than one might wish)
-        satisfied = Type.FromSameHead_Subtype(t, u, out var a, out var b);
-        if (satisfied) {
-          Contract.Assert(a.TypeArgs.Count == b.TypeArgs.Count);
-          var cl = a is UserDefinedType ? ((UserDefinedType)a).ResolvedClass : null;
-          for (int i = 0; i < a.TypeArgs.Count; i++) {
-            resolver.AllXConstraints.Add(new Resolver.XConstraint_EquatableArg(tok,
-              a.TypeArgs[i], b.TypeArgs[i],
-              a is CollectionType || (cl != null && cl.TypeArgs[i].Variance != TypeParameter.TPVariance.Non),
-              false,
-              errorMsg));
-            moreXConstraints = true;
+          // okay if t<:u or u<:t (this makes type inference more manageable, though it is more liberal than one might wish)
+          satisfied = Type.FromSameHead_Subtype(t, u, out var a, out var b);
+          if (satisfied) {
+            Contract.Assert(a.TypeArgs.Count == b.TypeArgs.Count);
+            var cl = a is UserDefinedType ? ((UserDefinedType)a).ResolvedClass : null;
+            for (int i = 0; i < a.TypeArgs.Count; i++) {
+              resolver.AllXConstraints.Add(new XConstraint_EquatableArg(tok,
+                a.TypeArgs[i], b.TypeArgs[i],
+                a is CollectionType || (cl != null && cl.TypeArgs[i].Variance != TypeParameter.TPVariance.Non),
+                false,
+                errorMsg));
+              moreXConstraints = true;
+            }
           }
+          break;
         }
-        break;
-      }
       case "Freshable": {
-        var collType = t.AsCollectionType;
-        if (collType is SetType || collType is SeqType) {
-          t = collType.Arg.NormalizeExpand();
+          var collType = t.AsCollectionType;
+          if (collType is SetType || collType is SeqType) {
+            t = collType.Arg.NormalizeExpand();
+          }
+          if (t is TypeProxy) {
+            return false;  // there is not enough information
+          }
+          satisfied = t.IsRefType;
+          break;
         }
-        if (t is TypeProxy) {
-          return false;  // there is not enough information
-        }
-        satisfied = t.IsRefType;
-        break;
-      }
       case "ModifiesFrame": {
-        var u = Types[1].NormalizeExpand();  // eventual ref type
-        var collType = t is MapType ? null : t.AsCollectionType;
-        if (collType != null) {
-          t = collType.Arg.NormalizeExpand();
-        }
-        if (t is TypeProxy) {
+          var u = Types[1].NormalizeExpand();  // eventual ref type
+          var collType = t is MapType ? null : t.AsCollectionType;
           if (collType != null) {
-            // we know enough to convert into a subtyping constraint
-            resolver.AddXConstraint(Token.NoToken/*bogus, but it seems this token would be used only when integers are involved*/, "IsRefType", t, errorMsg);
-            moreXConstraints = true;
+            t = collType.Arg.NormalizeExpand();
+          }
+          if (t is TypeProxy) {
+            if (collType != null) {
+              // we know enough to convert into a subtyping constraint
+              resolver.AddXConstraint(Token.NoToken/*bogus, but it seems this token would be used only when integers are involved*/, "IsRefType", t, errorMsg);
+              moreXConstraints = true;
+              resolver.ConstrainSubtypeRelation_Equal(u, t, errorMsg);
+              moreXConstraints = true;
+              convertedIntoOtherTypeConstraints = true;
+              return true;
+            } else {
+              return false;  // there is not enough information
+            }
+          }
+          if (t.IsRefType) {
             resolver.ConstrainSubtypeRelation_Equal(u, t, errorMsg);
-            moreXConstraints = true;
             convertedIntoOtherTypeConstraints = true;
             return true;
-          } else {
-            return false;  // there is not enough information
           }
+          satisfied = false;
+          break;
         }
-        if (t.IsRefType) {
-          resolver.ConstrainSubtypeRelation_Equal(u, t, errorMsg);
-          convertedIntoOtherTypeConstraints = true;
-          return true;
-        }
-        satisfied = false;
-        break;
-      }
       case "ReadsFrame": {
-        var u = Types[1].NormalizeExpand();  // eventual ref type
-        var arrTy = t.AsArrowType;
-        if (arrTy != null) {
-          t = arrTy.Result.NormalizeExpand();
-        }
-        var collType = t is MapType ? null : t.AsCollectionType;
-        if (collType != null) {
-          t = collType.Arg.NormalizeExpand();
-        }
-        if (t is TypeProxy) {
+          var u = Types[1].NormalizeExpand();  // eventual ref type
+          var arrTy = t.AsArrowType;
+          if (arrTy != null) {
+            t = arrTy.Result.NormalizeExpand();
+          }
+          var collType = t is MapType ? null : t.AsCollectionType;
           if (collType != null) {
-            // we know enough to convert into a subtyping constraint
-            resolver.AddXConstraint(Token.NoToken/*bogus, but it seems this token would be used only when integers are involved*/, "IsRefType", t, errorMsg);
+            t = collType.Arg.NormalizeExpand();
+          }
+          if (t is TypeProxy) {
+            if (collType != null) {
+              // we know enough to convert into a subtyping constraint
+              resolver.AddXConstraint(Token.NoToken/*bogus, but it seems this token would be used only when integers are involved*/, "IsRefType", t, errorMsg);
+              resolver.ConstrainSubtypeRelation_Equal(u, t, errorMsg);
+              moreXConstraints = true;
+              convertedIntoOtherTypeConstraints = true;
+              return true;
+            } else {
+              return false;  // there is not enough information
+            }
+          }
+          if (t.IsRefType && (arrTy == null || collType != null)) {
             resolver.ConstrainSubtypeRelation_Equal(u, t, errorMsg);
-            moreXConstraints = true;
             convertedIntoOtherTypeConstraints = true;
             return true;
-          } else {
-            return false;  // there is not enough information
           }
+          satisfied = false;
+          break;
         }
-        if (t.IsRefType && (arrTy == null || collType != null)) {
-          resolver.ConstrainSubtypeRelation_Equal(u, t, errorMsg);
-          convertedIntoOtherTypeConstraints = true;
-          return true;
-        }
-        satisfied = false;
-        break;
-      }
       default:
         Contract.Assume(false);  // unknown XConstraint
         return false;  // to please the compiler
@@ -557,5 +557,32 @@ public class XConstraint {
       }
     }
     return null;
+  }
+}
+
+public class XConstraintWithExprs : XConstraint {
+  public readonly Expression[] Exprs;
+  public XConstraintWithExprs(IToken tok, string constraintName, Type[] types, Expression[] exprs, TypeConstraint.ErrorMsg errMsg)
+    : base(tok, constraintName, types, errMsg) {
+    Contract.Requires(tok != null);
+    Contract.Requires(constraintName != null);
+    Contract.Requires(types != null);
+    Contract.Requires(exprs != null);
+    Contract.Requires(errMsg != null);
+    this.Exprs = exprs;
+  }
+}
+
+public class XConstraint_EquatableArg : XConstraint {
+  public bool AllowSuperSub;
+  public bool TreatTypeParamAsWild;
+  public XConstraint_EquatableArg(IToken tok, Type a, Type b, bool allowSuperSub, bool treatTypeParamAsWild, TypeConstraint.ErrorMsg errMsg)
+    : base(tok, "EquatableArg", new Type[] { a, b }, errMsg) {
+    Contract.Requires(tok != null);
+    Contract.Requires(a != null);
+    Contract.Requires(b != null);
+    Contract.Requires(errMsg != null);
+    AllowSuperSub = allowSuperSub;
+    TreatTypeParamAsWild = treatTypeParamAsWild;
   }
 }
